src/mcp-server/requirements.txt
fastmcp>=2.8.0
requests
pydantic
